Fix linux-specific clientaccess test
(cherry picked from commit 0d15457c7749efb8c71597491d42c00cf2c062b4)
