+++ pkg/clientaccess/token_test.go
@@ -24,60 +24,6 @@ var (
 	defaultToken    = "abcdef.0123456789abcdef"
 )
 
-// Test_UnitTrustedCA confirms that tokens are validated when the server uses a cert (self-signed or otherwise)
-// that is trusted by the OS CA bundle. This test must be run first, since it mucks with the system root certs.
-func Test_UnitTrustedCA(t *testing.T) {
-	assert := assert.New(t)
-	server := newTLSServer(t, defaultUsername, defaultPassword, false)
-	defer server.Close()
-	digest, _ := hashCA(getServerCA(server))
-
-	testInfo := &Info{
-		CACerts:  getServerCA(server),
-		BaseURL:  server.URL,
-		Username: defaultUsername,
-		Password: defaultPassword,
-		caHash:   digest,
-	}
-
-	testCases := []struct {
-		token    string
-		expected string
-	}{
-		{defaultPassword, ""},
-		{testInfo.String(), testInfo.Username},
-	}
-
-	// Point OS CA bundle at this test's CA cert to simulate a trusted CA cert.
-	// Note that this only works if the OS CA bundle has not yet been loaded in this process,
-	// as it is cached for the duration of the process lifetime.
-	// Ref: https://github.com/golang/go/issues/41888
-	path := t.TempDir() + "/ca.crt"
-	writeServerCA(server, path)
-	os.Setenv("SSL_CERT_FILE", path)
-
-	for _, testCase := range testCases {
-		info, err := ParseAndValidateToken(server.URL, testCase.token)
-		if assert.NoError(err, testCase) {
-			assert.Nil(info.CACerts, testCase)
-			assert.Equal(testCase.expected, info.Username, testCase.token)
-		}
-
-		info, err = ParseAndValidateToken(server.URL, testCase.token, WithUser("agent"))
-		if assert.NoError(err, testCase) {
-			assert.Nil(info.CACerts, testCase)
-			assert.Equal("agent", info.Username, testCase)
-		}
-	}
-
-	// Confirm that the cert is actually trusted by the OS CA bundle by making a request
-	// with empty cert pool
-	testInfo.CACerts = nil
-	res, err := testInfo.Get("/v1-k3s/server-bootstrap")
-	assert.NoError(err)
-	assert.NotEmpty(res)
-}
-
 // Test_UnitUntrustedCA confirms that tokens are validated when the server uses a self-signed cert
 // that is NOT trusted by the OS CA bundle.
 func Test_UnitUntrustedCA(t *testing.T) {
